Respond to review comments
